Reorganize Arch module
