+++ Modules/Kernel/Arch/CMakeLists.txt
@@ -1,20 +1,11 @@
-project(Loader)
+project(Arch)
 message(STATUS "    [-] Arch")
 
-add_library(Arch OBJECT
-    Arch.cpp
-    GDT.asm
-    GDT.cpp
-    IDT.asm
-    IDT.cpp
-    Interrupts.cpp
-    ISR.asm
-    PIC.cpp
-    Ports.cpp
+# Define an Arch target that must exist, that each individual architecture adds sources to
+add_library(Arch OBJECT)
+# Define the public CPU architecture API
+target_include_directories(Arch
+    PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
 
-# TODO: I have yet to decide if / how I want to go about public v. private headers
-#       in the kernel. For the Loader it makes more sense to keep them separate since
-#       the kernel and loader should be two distinct things that interact as little as
-#       possible (and mostly unidirectionally), but internal kernel code is different.
-target_include_directories(Arch PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+add_subdirectory(amd64)
