+++ Modules/Kernel/Arch/amd64/Devices/PIC.cpp
@@ -10,7 +10,7 @@
  */
 
 #include "PIC.hpp"
-#include "Ports.hpp"
+#include "IO/Ports.hpp"
 
 namespace PIC
 {
